backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.9
pgvector==0.4.1
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
insightface==0.7.3; platform_system != "Windows"
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
numpy==2.2.6
pillow==11.3.0
scikit-learn==1.7.1
tenacity==9.1.2
pytest==8.4.2
pytest-asyncio==1.1.0
